Data,Notice controller

+++ src/main/java/com/paytus/backend/Controller/NoticeController.java
@@ -1,38 +1,61 @@
 package com.paytus.backend.Controller;
 
+
 import com.paytus.backend.dto.NoticeDTO;
+import com.paytus.backend.model.response.CommonResult;
+import com.paytus.backend.model.response.ListResult;
+import com.paytus.backend.model.response.SingleResult;
 import com.paytus.backend.service.NoticeService;
+import com.paytus.backend.service.ResponseService;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+
 @RestController
 @RequestMapping("/api/notice")
 public class NoticeController {
     private NoticeService service;
-    public NoticeController(NoticeService service){
+
+    private final ResponseService responseService;
+    public NoticeController(NoticeService service, ResponseService responseService){
         this.service = service;
+        this.responseService = responseService;
     }
     @GetMapping("")
-    public List<NoticeDTO> getAllNotice() throws Exception {
-        return service.get();
+    public ListResult<NoticeDTO> getAllNotice() throws Exception {
+        return responseService.getListResult(service.get());
     }
     @GetMapping("/{noticeseq}")
-    public NoticeDTO getNotice(@PathVariable("noticeseq") int noticeseq) throws Exception {
-        return service.get(noticeseq);
+    public SingleResult<NoticeDTO> getNotice(@PathVariable("Noticeseq") int noticeseq) throws Exception {
+        return responseService.getSingleResult(service.get(noticeseq));
     }
     @PostMapping("")
-    public void registerNotice(@RequestBody NoticeDTO noticeDTO) throws Exception {
-        service.register(noticeDTO);
-        //System.out.println(noticeDTO);
+    public SingleResult<NoticeDTO> registerNotice(@RequestBody NoticeDTO noticeDTO) throws Exception {
+        if (noticeDTO.getNoticeseq()==0){
+            service.register(noticeDTO);
+            return responseService.getSingleResult(noticeDTO);
+        }
+        else
+            return responseService.getFailSingleResult();
+
     }
-    // @RequestBody: HTTP 요청의 바디내용을 통째로 자바객체로 변환해서 매핑된 메소드 파라미터로 전달해준다.
     @PutMapping("/{noticeseq}")
-    public void modifyNotice(@PathVariable("noticeseq") int noticeseq, @RequestBody NoticeDTO noticeDTO) throws Exception {
-        noticeDTO.setNoticeseq(noticeseq);
-        service.modify(noticeDTO);
+    public SingleResult<NoticeDTO> modifyNotice(@PathVariable("noticeseq") int noticeseq, @RequestBody NoticeDTO noticeDTO) throws Exception {
+        if (getNotice(noticeseq).getData()!=null){
+            noticeDTO.setNoticeseq(noticeseq);
+            service.modify(noticeDTO);
+
+            return responseService.getSingleResult(noticeDTO);
+        }
+        else
+            return responseService.getFailSingleResult();
     }
     @DeleteMapping("/{noticeseq}")
-    public void removeNotice(@PathVariable("noticeseq") int noticeseq) throws Exception {
-        service.remove(noticeseq);
+    public CommonResult removeNotice(@PathVariable("noticeseq") int noticeseq) throws Exception {
+        if (getNotice(noticeseq).getData()!=null){
+            service.remove(noticeseq);
+            return responseService.getSuccessResult();
+        }
+        else
+            return responseService.getFailResult();
     }
 }
